backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain-ollama==0.1.0
python-dotenv==1.0.0
fastapi-cors==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
email-validator==2.2.0
